requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.3.2
environs[django]==9.3.2
requests~=2.31.0
djangorestframework~=3.14.0
phonenumbers==8.13.20
dj-database-url==2.1.0
